==> 01-picasso_data.ipynb <==
# Picasso

import pandas as pd

## Loading data

picasso_df = pd.read_excel('data/src/Picasso_2000-2018.xlsx')
picasso_df.fillna('', inplace=True)

picasso_df.shape

picasso_df.head()

## Exploring data

### Unique values for each column

for c in picasso_df.columns:
    n = picasso_df[c].nunique()
    print(c, n)
    if n < 10:
        print(picasso_df[c].unique())

### Same values for 'xplace_t', 'xplace_r', and 'xplace_c'

a = b = c = 0
for i, row in picasso_df.drop_duplicates('ID')[['place_t' ,'place_r' ,'place_c']].iterrows():
    if row['place_t'] == row['place_r'] == row['place_c']:
        a += 1
    else:
        b += 1
    c += 1

a

b

c

a = b = c = 0
for i, row in picasso_df.drop_duplicates('xid')[['xpost_type', 'xplace_t' ,'xplace_r' ,'xplace_c']].iterrows():
    if row['xplace_t'] == row['xplace_r'] == row['xplace_c']:
        a += 1
    else:
        b += 1
    c += 1

a

b

c

### Values for 'xpost_type'

picasso_df[['xkey', 'xpost_type']].drop_duplicates().sort_values('xkey')

len(picasso_df[(picasso_df.xkey == 'coleccionista prestatario de obras') & (picasso_df.xpost_type == '')])

len(picasso_df[(picasso_df.xkey == 'coleccionista prestatario de obras') & (picasso_df.xpost_type == 'actor')])

len(picasso_df[(picasso_df.xkey == 'coleccionista prestatario de obras') & (picasso_df.xpost_type == 'entidad')])

## Completing 'xpost_type'

for i, row in picasso_df[picasso_df.xpost_type == ''].iterrows():
    if row.xkey == 'actor que participa como artista':
        picasso_df.loc[i, 'xpost_type'] = 'actor'
    elif row.xkey == 'coleccionista prestatario de obras':
        picasso_df.loc[i, 'xpost_type'] = 'entidad'
    elif row.xkey == 'entidad organizadora':
        picasso_df.loc[i, 'xpost_type'] = 'entidad'
    elif row.xkey == 'exposición de la que depende':
        picasso_df.loc[i, 'xpost_type'] = 'exposición'
    elif row.xkey == 'fuente de información':
        picasso_df.loc[i, 'xpost_type'] = 'entidad'

picasso_df[['xkey', 'xpost_type']].drop_duplicates().sort_values('xkey')

## Stats

picasso_df.xkey.value_counts()

picasso_df.drop_duplicates('xid').xpost_type.value_counts()

a = b = c = 0
exhibition_ids = picasso_df.ID.unique()
for x in picasso_df[picasso_df.xpost_type == 'exposición'].xid.unique():
    if x in exhibition_ids:
        a += 1
    else:
        b += 1
    c += 1

a

b

c

## Saving data

picasso_df.to_csv('data/out/picasso.csv', index=False)
