==> 02-cleaning_data.ipynb <==
# Exhibitions

from collections import Counter

import pandas as pd

## Loading data

%%time

exhibitions_df = pd.read_excel('data/src/20180410_EXHIBITIONS.xlsx')
exhibitions_df.fillna('', inplace=True)

exhibitions_df.shape

exhibitions_df.head()

## Cleaning data

### Types

for c in exhibitions_df.columns:
    print(c, exhibitions_df[c].dtype)

### Blank values in important columns

important_columns = ['ID', 'post_type', 'post_title', 'xkey', 'xid', 'xpost_type', 'xpost_title']

for c in important_columns:
    counts = exhibitions_df[c].value_counts()
    if '' in counts:
        print(c, counts[''])

### 'xkey' and 'xid'

exhibitions_df[exhibitions_df.xkey == ''][important_columns]

for ide in exhibitions_df[exhibitions_df.xkey == ''].ID:
    print(ide, exhibitions_df.ID.value_counts()[ide])

exhibitions_df = exhibitions_df[exhibitions_df.xkey != '']

exhibitions_df['xid'] = exhibitions_df.xid.astype(int)

### 'xpost_type'

exhibitions_df[['xkey', 'xpost_type']].drop_duplicates().sort_values('xkey')

for i, row in exhibitions_df[exhibitions_df.xpost_type == ''].iterrows():
    if row.xkey == 'actor comisario':
        exhibitions_df.loc[i, 'xpost_type'] = 'actor'
    elif row.xkey == 'actor que participa como artista':
        exhibitions_df.loc[i, 'xpost_type'] = 'actor'
    elif row.xkey == 'catálogo':
        exhibitions_df.loc[i, 'xpost_type'] = 'catálogo'
    elif row.xkey == 'coleccionista prestatario de obras':
        exhibitions_df.loc[i, 'xpost_type'] = 'entidad'
    elif row.xkey == 'empresa que realiza la museografía':
        exhibitions_df.loc[i, 'xpost_type'] = 'empresa'
    elif row.xkey == 'entidad organizadora':
        exhibitions_df.loc[i, 'xpost_type'] = 'entidad'
    elif row.xkey == 'entidad patrocinadora':
        exhibitions_df.loc[i, 'xpost_type'] = 'entidad'
    elif row.xkey == 'exposición de la que depende':
        exhibitions_df.loc[i, 'xpost_type'] = 'exposición'
    elif row.xkey == 'fuente de información':
        exhibitions_df.loc[i, 'xpost_type'] = 'entidad'

for i, row in exhibitions_df[
    (exhibitions_df.xkey == 'actor que participa como artista') &
    (exhibitions_df.xpost_type != 'actor')
    ].iterrows():
        exhibitions_df.loc[i, 'xpost_type'] = 'actor'

### Duplicated 'xid' and 'xpost_type'

exhibitions_df[['xid', 'xpost_type']].drop_duplicates().xid.value_counts().head()

exhibitions_df[exhibitions_df.xid == -1][important_columns]

exhibitions_df.loc[9349, 'xid'] = 1
exhibitions_df.loc[72626, 'xid'] = 2

exhibitions_df[exhibitions_df.xid == 28487][important_columns]

exhibitions_df.loc[40321, 'xpost_type'] = 'actor'

exhibitions_df[exhibitions_df.xid == 743].drop_duplicates(['xid', 'xpost_type'])[important_columns]

exhibitions_df[(exhibitions_df.xid == 743) & (exhibitions_df.xpost_type != 'entidad')][important_columns]

exhibitions_df = exhibitions_df.drop([29974, 67917], axis=0)

### 'xpost_title'

for i, row in exhibitions_df[exhibitions_df.xpost_title == ''].iterrows():
    exhibitions_df.loc[i, 'xpost_title'] = 'Desconocido'

### Gender

exhibitions_df[exhibitions_df.xpost_type != 'actor'].xgender.value_counts()

exhibitions_df[exhibitions_df.xpost_type == 'actor'].xgender.value_counts()

%%time

for i, row in exhibitions_df[
    (exhibitions_df.xpost_type == 'actor') &
    (exhibitions_df.xgender == '')
    ].iterrows():
        exhibitions_df.loc[i, 'xgender'] = 'No declarado'

for i, row in exhibitions_df[
    (exhibitions_df.xpost_type != 'actor') &
    (exhibitions_df.xgender == '')
    ].iterrows():
        exhibitions_df.loc[i, 'xgender'] = 'NO APLICA'

## Testing data

### Types

exhibitions_df.xid.dtype

### Blank values in important columns

for c in important_columns:
    counts = exhibitions_df[c].value_counts()
    if '' in counts:
        print(c, counts[''])

### 'xkey' and 'xid'

exhibitions_df[exhibitions_df.xkey == '']

### 'xpost_type'

exhibitions_df[['xkey', 'xpost_type']].drop_duplicates().sort_values('xkey')

### Duplicated 'xid' and 'xpost_type'

exhibitions_df[['xid', 'xpost_type']].drop_duplicates().xid.value_counts().head()

### 'xpost_title'

exhibitions_df[exhibitions_df.xpost_title == '']

### Gender

exhibitions_df.xgender.value_counts()

### x-data consistency

x_columns = [c for c in exhibitions_df.columns if c.startswith('x')]
x_columns.remove('xkey')

%%time

for ide in exhibitions_df.xid.unique():
    s = set(tuple(row) for i, row in exhibitions_df[exhibitions_df.xid == ide][x_columns].iterrows())
    assert(len(s) == 1)

## Saving data

exhibitions_df.to_csv('data/out/exhibitions.csv', index=False)
