==> 03-exploring_data.ipynb <==
# Exhibitions

from collections import Counter

import pandas as pd

## Loading data

%%time

exhibitions_df = pd.read_csv('data/out/exhibitions.csv')
exhibitions_df.fillna('', inplace=True)

exhibitions_df.shape

exhibitions_df.head()

## Exploring data

important_columns = ['ID', 'post_type', 'post_title', 'xkey', 'xid', 'xpost_type', 'xpost_title']

### Unique values for important columns

for c in important_columns:
    n = exhibitions_df[c].nunique()
    print(c, n)
    if n <= 10:
        print(exhibitions_df[c].unique())

### Total number of unique elements

len(set(exhibitions_df.ID.unique()).union(exhibitions_df.xid.unique()))

### Number of unique exhibitions

exhibitions_df.ID.nunique()

### Number of unique related elements

exhibitions_df.xid.nunique()

### Number of unique elements for each type

exhibitions_df.drop_duplicates('xid').xpost_type.value_counts()

### Total number of relationships

exhibitions_df.xkey.value_counts()

### Types of relationships

exhibitions_df[['xkey', 'xpost_type']].drop_duplicates().sort_values('xkey')

### Number of unique actors and genders

exhibitions_df[exhibitions_df.xpost_type == 'actor'].xid.nunique()

exhibitions_df[exhibitions_df.xpost_type == 'actor'].drop_duplicates('xid').xgender.value_counts()

### Same values for '(x)place_t', '(x)place_r', and '(x)place_c'

#### Same values for 'place_t', 'place_r', and 'place_c'

a = b = c = 0
for i, row in exhibitions_df.drop_duplicates('ID').iterrows():
    if row['place_t'] == row['place_r'] == row['place_c']:
        a += 1
    else:
        b += 1
    c += 1

a

b

c

#### Same values for 'xplace_t', 'xplace_r', and 'xplace_c'

a = b = c = 0
correct_places = []
incorrect_places = []
for i, row in exhibitions_df.drop_duplicates('xid').iterrows():
    if row['xplace_t'] == row['xplace_r'] == row['xplace_c']:
        a += 1
        incorrect_places.append(row['xpost_type'])
    else:
        b += 1
        correct_places.append(row['xpost_type'])
    c += 1

c

a

Counter(incorrect_places).most_common()

b

Counter(correct_places).most_common()

### Self-contained exhibitions

a = b = c = 0
exhibition_ids = exhibitions_df.ID.unique()
for x in exhibitions_df[exhibitions_df.xpost_type == 'exposición'].xid.unique():
    if x in exhibition_ids:
        a += 1
    else:
        b += 1
    c += 1

a

b

c
