==> 04-generating_graph.ipynb <==
# Exhibitions

import networkx as nx
import pandas as pd

## Loading data

%%time

exhibitions_df = pd.read_csv('data/out/exhibitions.csv')
exhibitions_df.fillna('', inplace=True)

exhibitions_df.shape

exhibitions_df.head()

## Exploring values by type

for t in exhibitions_df.xpost_type.unique():
    print(t)
    print(exhibitions_df[exhibitions_df.xpost_type == t].xpost_title.value_counts()[:5])
    print()

exhibitions_df[exhibitions_df.xpost_title == '[Sin determinar]'].xid.unique()

exhibitions_df[exhibitions_df.xpost_title == 'Colectivo sin determinar'].xid.unique()

## Processing data

def extract_e_info(row):
    return {
        'id': row.ID,
        'type': row.post_type,
        'title': row.post_title,
        'place_t': row.place_t,
        'place_r': row.place_r,
        'place_c': row.place_c,
        'start_y': row.start_y,
        'start_m': row.start_m,
        'start_d': row.start_d,
        'end_y': row.end_y,
        'end_m': row.end_m,
        'end_d': row.end_d,
    }

def extract_x_info(row):
    return {
        'id': row.xid,
        'type': row.xpost_type,
        'title': row.xpost_title,
        'place_t': row.xplace_t,
        'place_r': row.xplace_r,
        'place_c': row.xplace_c,
        'start_y': row.xstart_y,
        'start_m': row.xstart_m,
        'start_d': row.xstart_d,
        'end_y': row.xend_y,
        'end_m': row.xend_m,
        'end_d': row.xend_d,
        'gender': row.xgender,
    }

%%time

g = nx.DiGraph()
for i, row in exhibitions_df.iterrows():
    e = extract_e_info(row)
    x = extract_x_info(row)
    g.add_node(e['id'], **e)
    g.add_node(x['id'], **x)
    g.add_edge(e['id'], x['id'], label=row.xkey)

g.remove_node(55293) # Removing actor '[Sin determinar]'
g.remove_node(23812) # Removing actor 'Colectivo sin determinar'
for n, data in list(g.nodes(data=True)): # Removing nodes of type 'catálogo' and 'empresa'
    if data['type'] == 'catálogo' or data['type'] == 'empresa':
        g.remove_node(n)

g.number_of_nodes()

g.number_of_edges()

g.number_of_selfloops()

nx.write_gexf(g, 'data/out/exhibitions.gexf')
