==> 05-generating_histogram.ipynb <==
# Exhibitions

from collections import Counter
from pprint import pprint
from string import punctuation

from nltk import word_tokenize
from nltk.corpus import stopwords
import pandas as pd

## Loading data

%%time

exhibitions_df = pd.read_csv('data/out/exhibitions.csv')
exhibitions_df.fillna('', inplace=True)

exhibitions_df.shape

exhibitions_df.head()

## Processing data

def add_blank(x):
    return x + ' - '

exhibitions_df.post_title = exhibitions_df.post_title.apply(add_blank)

exhibitions_df.head()

stop = stopwords.words('english') + stopwords.words('spanish') + list(punctuation)

def clean(s):
    r = s.lower().strip()
    rs = [w for w in word_tokenize(r) if w not in stop]
    r = ' '.join(rs)
    for x in list("'`“”‘’¡¿…–"):
        r = r.replace(x, '')
    r = ' '.join(r.split())
    return r

### General histogram

all_text = exhibitions_df.drop_duplicates('ID').post_title.sum()

all_text[:100]

cleaned_all_text = clean(all_text)

cleaned_all_text[:100]

general_counter = Counter(cleaned_all_text.split())

general_counter.most_common(20)

### Histogram by city

sp_exhibitions = exhibitions_df[exhibitions_df.place_c == 'España'].drop_duplicates('ID')

sp_exhibitions.place_t.value_counts()[:10]

top_sp_cities = [t for t, f in sp_exhibitions.place_t.value_counts()[:10].iteritems()]

top_sp_cities

top_sp_cities_exhibitions = sp_exhibitions[sp_exhibitions.place_t.isin(top_sp_cities)]

text_by_city_sr = top_sp_cities_exhibitions.groupby('place_t').post_title.sum()

for city in top_sp_cities:
    print(city)
    print('-' * 80)
    text = text_by_city_sr[city]
    cleaned_text = clean(text)
    counter = Counter(cleaned_text.split())
    pprint(counter.most_common(20))
    print('=' * 80)
    print()
